requirements.txt
asgiref==3.8.1
attrs==25.3.0
bcrypt==4.2.0
black==24.10.0
boto3==1.35.29
botocore==1.35.56
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
django==5.1.4
django-cors-headers==4.7.0
django-db-connection-pool==1.2.5
django-db-geventpool==4.0.7
django-debug-toolbar==5.0.1
django-extensions==3.2.3
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0
et-xmlfile==2.0.0
fonttools==4.56.0
greenlet==3.1.1
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
matplotlib==3.10.0
mypy-extensions==1.0.0
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==10.4.0
pip==24.0
platformdirs==4.3.6
plotly==5.24.1
psycopg==3.2.3
psycopg-pool==3.2.4
psycopg2-binary==2.9.9
pyjwt==2.9.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyyaml==6.0.2
referencing==0.36.2
requests==2.28.2
rpds-py==0.23.1
s3transfer==0.10.3
six==1.16.0
sqlalchemy==2.0.36
sqlparams==6.1.0
sqlparse==0.5.1
tenacity==9.0.0
typing-extensions==4.12.2
tzdata==2024.1
unidecode==1.3.8
uritemplate==4.1.1
urllib3==1.26.15
uv==0.6.6
whitenoise==6.8.2
pre-commit==4.2.0
qrcode==8.0
gunicorn==23.0.0
pytest==8.2.0
pytest-django==4.8.0
